feat: Implemented Game Engine and postgres docker compose startup
Populated the docker compose for game-engine and postgres which will be where we start. We confirmed we can hit /ping on the game engine locally and we confirmed the game engine can connect to the postgres database. Now we need to set up the infrastructure to set up the DB and then make sure game-engine can do CRUD operations on postgres db.