feat: Implemented the migrations themselves
Implemented the migrations themselves to occur at startup, now when the docker container starts up the first time, the migration files are sequentially ran so populate the database. as we need to add more characters/data we can just keep adding more migration tables. The logic that checks if migrations were already done was also tested and is successful.