feat: Implementing code for future pages (currently drawn out)
Implemented the info and roster pages which will be the first to be developed. The pages have been drawn out, i will just first go to the game-engine and get the API calls / infrastructure changes necessary for some of the features i want to be able to call from the website and from the discord bot as well since development on that is starting soon.