feat: Implemented ability gating infrastructure,
We now check the abilities prereq table to make sure that a character can learn this ability based off of the prereqs. We still need to add the check that sees if the character can learn the first ability in the first place (character_abilities). Added logic that parses the prereqs tree, will be useful in website as well. Added endpoints to check details for an ability and also for its prereqs. Added serializers for ability and ability prereqs, new "No content" response and constant,