feat: Infrastructure impl of the enemy (bot) feature allowing users to fight against other users and also other bots
Implemented two tables, enemy, which will track all the enemy specific info and enemy_abilities, a junction table tracking which enemies have access to which abilities. enemies will be kept separate from claimable characters for now. also added the repository and service which will be used later  and added them in the main to be used already